fix(anvil): respect historic logs in newFilter
